wip: working on buy token screen
