+++ apps/wallet/features/pool/sections/poolActions.tsx
@@ -1,13 +1,11 @@
-import { FC, useMemo } from "react";
-import { useAccount } from "@/app/hooks/useAccount";
-import { Number } from "@/app/components/number";
-import { useRouter } from "next/router";
+import { FC } from "react";
 import { PoolContractData } from "@/types/poolContractData";
 // @ts-ignore
 import hashicon from "hashicon";
 import { useTranslation } from "next-i18next";
 import { Button } from "react-daisyui";
 import { openExternalUrl } from "@/app/openExternalUrl";
+import { useRouter } from "next/router";
 
 interface Props {
   poolData: PoolContractData;
@@ -16,20 +14,23 @@ interface Props {
 export const PoolActions: FC<Props> = ({ poolData }) => {
   // const account = useAccount();
   const { t } = useTranslation();
+  const router = useRouter();
 
   const handleOpenDocumentation = () => {
     const poolId = poolData.poolId;
     openExternalUrl("");
   };
 
-  const handleAcquireShare = () => {};
+  const handleAcquireShare = async () => {
+    await router.push(`/pool/${poolData.poolId}/acquisition`);
+  };
 
   return (
     <div className="p-2 flex-row flex mx-auto justify-center">
       <Button color="primary" onClick={handleAcquireShare}>
         {t("buy_token")}
       </Button>
-      <Button className="ml-4" onClick={handleOpenDocumentation}>
+      <Button className="ml-4" color="accent" onClick={handleOpenDocumentation}>
         {t("show_doc")}
       </Button>
     </div>
