+++ apps/wallet/package.json
@@ -45,6 +45,7 @@
     "react-device-detect": "^2.2.2",
     "react-dom": "^18.2.0",
     "react-icons": "^4.4.0",
+    "react-qr-code": "^2.0.8",
     "react-redux": "^8.0.2",
     "redux-persist": "^6.0.0",
     "swr": "^1.3.0"
@@ -63,6 +64,7 @@
     "lint-staged": "^13.0.3",
     "postcss": "^8.4.14",
     "prettier": "2.7.1",
+    "tailwind-scrollbar": "^2.0.1",
     "tailwindcss": "^3.1.7",
     "ts-jest": "^28.0.7",
     "typescript": "4.7.4"
