+++ apps/wallet/tailwind.config.js
@@ -42,5 +42,12 @@ module.exports = {
     prefix: "",
     darkTheme: "luxury",
   },
-  plugins: [require("@tailwindcss/typography"), require("daisyui")],
+  plugins: [
+    require("@tailwindcss/typography"),
+    require("daisyui"),
+    require("tailwind-scrollbar"),
+  ],
+  variants: {
+    scrollbar: ["rounded"],
+  },
 };
